Convert Long to Integer for Database Compatibility

+++ flower/src/main/java/com/flower/config/jwt/UserDetailsImpl.java
@@ -49,7 +49,7 @@ public class UserDetailsImpl implements UserDetails {
         return true;
     }
 
-    public Long getUserId() {
+    public Integer getUserId() {
         return user.getUserId();
     }
