+++ flower/src/main/java/com/flower/repository/FlowerRepository.java
@@ -9,7 +9,7 @@ import java.time.LocalDate;
 import java.util.List;
 import java.util.Optional;
 
-public interface FlowerRepository extends JpaRepository<Flower, Long> {
+public interface FlowerRepository extends JpaRepository<Flower, Integer> {
 
     // 오늘의 꽃 조회 (월일만 비교, 년도 무시)
     @Query(value = "SELECT * FROM Flower WHERE MONTH(Today_Flower) = :month AND DAY(Today_Flower) = :day LIMIT 1", nativeQuery = true)
